refactor(workflow): inline provider selector sub-blocks
