+++ apps/tradinggoose/widgets/widgets/editor_workflow/components/workflow-block/components/sub-block/sub-block.tsx
@@ -2,7 +2,10 @@ import type React from 'react'
 import { memo, useEffect, useMemo, useRef, useState } from 'react'
 import { format } from 'date-fns'
 import { AlertTriangle, Info } from 'lucide-react'
-import { Label, Tooltip, TooltipContent, TooltipTrigger } from '@/components/ui'
+import { MarketProviderSelector } from '@/components/market-selector/provider-selector'
+import { TradingAccountSelector } from '@/components/trading-selector/account-selector'
+import { TradingProviderSelector } from '@/components/trading-selector/provider-selector'
+import { Label, Tooltip, TooltipContent, TooltipProvider, TooltipTrigger } from '@/components/ui'
 import { DateTimePicker } from '@/components/ui/datetime-picker'
 import { SimpleTimePicker } from '@/components/ui/simple-time-picker'
 import { Slider } from '@/components/ui/slider'
@@ -15,6 +18,20 @@ import {
 } from '@/lib/time-format'
 import { cn } from '@/lib/utils'
 import type { SubBlockConfig } from '@/blocks/types'
+import { useOAuthProviderAvailability } from '@/hooks/queries/oauth-provider-availability'
+import {
+  getMarketProviderOptions,
+  getMarketProviderOptionsByKind,
+} from '@/providers/market/providers'
+import {
+  type PortfolioIdentity,
+  toPortfolioValueObject,
+} from '@/providers/trading/portfolio-identity'
+import {
+  getTradingWidgetProviderAvailabilityIds,
+  getTradingWidgetProviderOptions,
+} from '@/widgets/utils/trading-widget-providers'
+import { useOptionalWorkflowRoute } from '@/widgets/widgets/editor_workflow/context/workflow-route-context'
 import {
   ChannelSelectorInput,
   CheckboxList,
@@ -48,12 +65,10 @@ import {
   ToolInput,
   TriggerSave,
   VariablesInput,
-  WorkflowMarketProviderSelector,
-  WorkflowTradingAccountSelector,
-  WorkflowTradingProviderSelector,
 } from '@/widgets/widgets/editor_workflow/components/workflow-block/components/sub-block/components'
 import { DocumentTagEntry } from './components/document-tag-entry/document-tag-entry'
 import { KnowledgeTagFilters } from './components/knowledge-tag-filters/knowledge-tag-filters'
+import { useDependsOnGate } from './hooks/use-depends-on-gate'
 import { useSubBlockValue } from './hooks/use-sub-block-value'
 
 interface SubBlockProps {
@@ -240,6 +255,181 @@ function SubBlockDateTimeField({
   )
 }
 
+const readSelectorValue = (value: unknown) => {
+  if (typeof value === 'string') return value.trim()
+  if (value && typeof value === 'object' && 'value' in value) {
+    const nestedValue = (value as { value?: unknown }).value
+    return typeof nestedValue === 'string' ? nestedValue.trim() : ''
+  }
+  return ''
+}
+
+function useOptionValueSync(
+  value: string,
+  setValue: (value: string) => void,
+  options: Array<{ id: string }>,
+  disabled: boolean,
+  autoSelectFirstOption: boolean
+) {
+  useEffect(() => {
+    if (disabled) return
+
+    const optionIds = options.map((option) => option.id)
+    if (value && optionIds.includes(value)) return
+
+    const nextValue = autoSelectFirstOption ? (optionIds[0] ?? '') : ''
+    if (value !== nextValue) {
+      setValue(nextValue)
+    }
+  }, [autoSelectFirstOption, disabled, options, setValue, value])
+}
+
+function SubBlockMarketProviderSelector({
+  blockId,
+  config,
+  disabled,
+  contextValues,
+}: {
+  blockId: string
+  config: SubBlockConfig
+  disabled: boolean
+  contextValues?: Record<string, any>
+}) {
+  const [value, setValue] = useSubBlockValue<string>(blockId, config.id)
+  const { finalDisabled } = useDependsOnGate(blockId, config, { disabled, contextValues })
+  const options = useMemo(
+    () =>
+      config.marketProviderKind
+        ? getMarketProviderOptionsByKind(config.marketProviderKind)
+        : getMarketProviderOptions(),
+    [config.marketProviderKind]
+  )
+  const selectedValue = readSelectorValue(value)
+
+  useOptionValueSync(
+    selectedValue,
+    setValue,
+    options,
+    finalDisabled,
+    config.autoSelectFirstOption !== false
+  )
+
+  return (
+    <TooltipProvider>
+      <MarketProviderSelector
+        value={selectedValue}
+        options={options}
+        onChange={setValue}
+        disabled={finalDisabled}
+        placeholder={config.placeholder}
+        variant='form'
+      />
+    </TooltipProvider>
+  )
+}
+
+function SubBlockTradingProviderSelector({
+  blockId,
+  config,
+  disabled,
+  contextValues,
+}: {
+  blockId: string
+  config: SubBlockConfig
+  disabled: boolean
+  contextValues?: Record<string, any>
+}) {
+  const kind = config.tradingProviderKind ?? 'order'
+  const [value, setValue] = useSubBlockValue<string>(blockId, config.id)
+  const { finalDisabled } = useDependsOnGate(blockId, config, { disabled, contextValues })
+  const availabilityIds = useMemo(() => getTradingWidgetProviderAvailabilityIds(kind), [kind])
+  const availabilityQuery = useOAuthProviderAvailability(availabilityIds, !finalDisabled)
+  const options = useMemo(
+    () =>
+      availabilityQuery.data ? getTradingWidgetProviderOptions(kind, availabilityQuery.data) : [],
+    [kind, availabilityQuery.data]
+  )
+  const selectedValue = readSelectorValue(value)
+
+  useOptionValueSync(
+    selectedValue,
+    setValue,
+    options,
+    finalDisabled || availabilityQuery.isLoading,
+    config.autoSelectFirstOption !== false
+  )
+
+  return (
+    <TooltipProvider>
+      <TradingProviderSelector
+        value={selectedValue}
+        options={options}
+        onChange={setValue}
+        disabled={finalDisabled || availabilityQuery.isLoading}
+        placeholder={config.placeholder}
+        variant='form'
+      />
+    </TooltipProvider>
+  )
+}
+
+function SubBlockTradingAccountSelector({
+  blockId,
+  config,
+  disabled,
+  contextValues,
+}: {
+  blockId: string
+  config: SubBlockConfig
+  disabled: boolean
+  contextValues?: Record<string, any>
+}) {
+  const route = useOptionalWorkflowRoute()
+  const providerFieldId = config.tradingProviderFieldId ?? 'provider'
+  const [value, setValue] = useSubBlockValue<PortfolioIdentity | ''>(blockId, config.id)
+  const [storeProviderValue] = useSubBlockValue<string>(blockId, providerFieldId)
+  const [requestedServiceId, setRequestedServiceId] = useState<string | null>(null)
+  const { finalDisabled } = useDependsOnGate(blockId, config, { disabled, contextValues })
+  const providerId =
+    readSelectorValue(contextValues?.[providerFieldId]) || readSelectorValue(storeProviderValue)
+  const portfolioIdentity = useMemo(() => toPortfolioValueObject(value), [value])
+
+  useEffect(() => {
+    setRequestedServiceId(null)
+  }, [providerId])
+
+  useEffect(() => {
+    if (portfolioIdentity && portfolioIdentity.providerId !== providerId) {
+      setValue('')
+    }
+  }, [portfolioIdentity, providerId, setValue])
+
+  return (
+    <TooltipProvider>
+      <TradingAccountSelector
+        workspaceId={route?.workspaceId}
+        providerId={providerId}
+        serviceId={requestedServiceId ?? portfolioIdentity?.serviceId}
+        portfolioIdentity={portfolioIdentity}
+        disabled={finalDisabled}
+        placeholder={config.placeholder}
+        tooltipText={config.tooltip ?? config.description ?? 'Select trading account'}
+        toolName='Trading'
+        onAccountSelect={(selection) => {
+          setRequestedServiceId(
+            selection.serviceId ?? selection.portfolioIdentity?.serviceId ?? null
+          )
+          const nextIdentity = selection.portfolioIdentity
+            ? toPortfolioValueObject(selection.portfolioIdentity)
+            : null
+          setValue(nextIdentity ?? '')
+        }}
+        variant='form'
+      />
+    </TooltipProvider>
+  )
+}
+
 export const SubBlock = memo(
   function SubBlock({
     blockId,
@@ -417,9 +607,8 @@ export const SubBlock = memo(
           )
         case 'market-provider-selector':
           return (
-            <WorkflowMarketProviderSelector
+            <SubBlockMarketProviderSelector
               blockId={blockId}
-              subBlockId={config.id}
               disabled={isDisabled}
               config={config}
               contextValues={contextValues}
@@ -427,9 +616,8 @@ export const SubBlock = memo(
           )
         case 'trading-provider-selector':
           return (
-            <WorkflowTradingProviderSelector
+            <SubBlockTradingProviderSelector
               blockId={blockId}
-              subBlockId={config.id}
               disabled={isDisabled}
               config={config}
               contextValues={contextValues}
@@ -437,9 +625,8 @@ export const SubBlock = memo(
           )
         case 'trading-account-selector':
           return (
-            <WorkflowTradingAccountSelector
+            <SubBlockTradingAccountSelector
               blockId={blockId}
-              subBlockId={config.id}
               disabled={isDisabled}
               config={config}
               contextValues={contextValues}
